Use julia-snail
Also put some display stuff closer to the beginning of the config.
I'm noticing that pinentry sometimes fails in Emacs, and usually
it takes a restart of Emacs or the whole X server to get it working
again... that bit needs some work (I think it's probably a problem
with X configuration).